Modified index.html and code.gs to add an option for non-destructive translation.
If the user selects it, they can have the translation done in a new cell.  The function translate() was modified
to check the user's choice.  It sets a new variable for target sheet which is a copy of the active sheet with
appended text in the name.  It is then passed to the other functions.

These other functions now update to the target sheet instead of the active sheet.  Depending on the choice of
the user, the target sheet could be the active sheet or could be a new sheet.

The function which translates only selected text was modified to work on the original selected range, but on the
target sheet instead of the active sheet.  Also added colouring the translated range (should this be a user
option?) to help identify the specific cells translated.